# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AosAdjutant.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/factions/{factionId}/abilities": {
      "post": {
        "tags": [
          "Factions"
        ],
        "summary": "Create an ability for a faction",
        "parameters": [
          {
            "name": "factionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Factions"
        ],
        "summary": "Get all abilities for a faction",
        "parameters": [
          {
            "name": "factionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Phase",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Phase"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AbilitySortBy"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/battle-formations/{battleFormationId}/abilities": {
      "post": {
        "tags": [
          "Battle Formations"
        ],
        "summary": "Create an ability for a battle formation",
        "parameters": [
          {
            "name": "battleFormationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAbilityDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbilityResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Battle Formations"
        ],
        "summary": "Get all abilities for a battle formation",
        "parameters": [
          {
            "name": "battleFormationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Phase",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Phase"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/AbilitySortBy"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseOfAbilityResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbilityResponseDto": {
        "required": [
          "abilityId",
          "name",
          "reaction",
          "declaration",
          "effect",
          "phase",
          "restriction",
          "turn",
          "version"
        ],
        "type": "object",
        "properties": {
          "abilityId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "reaction": {
            "type": [
              "null",
              "string"
            ]
          },
          "declaration": {
            "type": [
              "null",
              "string"
            ]
          },
          "effect": {
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase"
          },
          "restriction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Restriction"
              }
            ]
          },
          "turn": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Turn"
              }
            ]
          },
          "version": {
            "type": "integer",
            "format": "uint32"
          }
        }
      },
      "AbilitySortBy": {
        "enum": [
          "Name",
          "Phase",
          null
        ]
      },
...
      "CreateAbilityDto": {
        "required": [
          "name",
          "effect",
          "phase"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "reaction": {
            "maxLength": 100,
            "minLength": 1,
            "type": [
              "null",
              "string"
            ]
          },
          "declaration": {
            "maxLength": 100,
            "minLength": 1,
            "type": [
              "null",
              "string"
            ]
          },
          "effect": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase"
          },
          "restriction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Restriction"
              }
            ]
          },
          "turn": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Turn"
              }
            ]
          }
        }
      },
...
      "PaginatedResponseOfAbilityResponseDto": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AbilityResponseDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Phase": {
        "enum": [
          "Deployment",
          "Start",
          "Hero",
          "Movement",
          "Shooting",
          "Charge",
          "Combat",
          "End",
          "Passive"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "Restriction": {
        "enum": [
          "OnceTurnArmy",
          "OnceRoundArmy",
          "OnceBattleArmy",
          "OnceRound",
          "OnceBattle",
          null
        ]
      },
      "SortDirection": {
        "enum": [
          "Asc",
          "Desc"
        ]
      },
      "Turn": {
        "enum": [
          "YourTurn",
          "EnemyTurn",
          "AnyTurn",
          null
        ]
      },
...
    }
  },
...
}